requirements.txt
Flask==3.0.0
psycopg2-binary==2.9.9
geopandas==0.14.3
shapely==2.0.3
pyproj==3.6.1
fiona==1.9.5
